Reverted changes

+++ 2-do_deploy_web_static.py
@@ -18,42 +18,51 @@ def do_deploy(archive_path):
 
     # print("type: {} | val = {}".format(type(archive_path), archive_path))
 
-    try:
-        # Get name for unzipped file
-        archive = archive_path.split(sep="versions/")[1]
-        unzipped = archive.split(sep=".tgz")[0]
-        # print("Unzipped: {}".format(unzipped))
+    # Get name for unzipped file
+    archive = archive_path.split(sep="versions/")[1]
+    unzipped = archive.split(sep=".tgz")[0]
+    # print("Unzipped: {}".format(unzipped))
 
-        # Upload archive
-        run("mkdir --parents /tmp/")
-        put(local_path=archive_path, remote_path="/tmp/")
+    # Upload archive
+    if run("mkdir --parents /tmp/").failed is True:
+        return False
 
-        # Uncompress archive
-        cmd = "mkdir --parents /data/web_static/releases/{}/".format(unzipped)
-        run(cmd)
+    if put(local_path=archive_path, remote_path="/tmp/").failed is True:
+        return False
 
-        cmd = "tar -xzf /tmp/{} -C /data/web_static/releases/{}/"
-        cmd = cmd.format(archive, unzipped)
-        run(cmd)
+    # Uncompress archive
+    cmd = "mkdir --parents /data/web_static/releases/{}/".format(unzipped)
+    if run(cmd).failed is True:
+        return False
 
-        # remove temporary file
-        cmd = "rm -rf /tmp/{}".format(archive)
-        run(cmd)
-        cmd = "mv -f /data/web_static/releases/{}/web_static/*".format(unzipped)
-        cmd = cmd + " /data/web_static/releases/{}/".format(unzipped)
-        run(cmd)
+    cmd = "tar -xzf /tmp/{} -C /data/web_static/releases/{}/"
+    cmd = cmd.format(archive, unzipped)
+    if run(cmd).failed is True:
+        return False
 
-        cmd = "rm -rf /data/web_static/releases/{}/web_static".format(unzipped)
-        run(cmd)
+    # remove temporary file
+    cmd = "rm --force /tmp/{}".format(archive)
+    if run(cmd).failed is True:
+        return False
 
-        # Relink deployed code
-        run("rm -rf /data/web_static/current")
-        cmd = "ln -s /data/web_static/releases/{}/".format(unzipped)
-        cmd = cmd + " /data/web_static/current"
-        run(cmd)
-    except Exception:
+    cmd = "mv -f /data/web_static/releases/{}/web_static/*".format(unzipped)
+    cmd = cmd + " /data/web_static/releases/{}/".format(unzipped)
+    if run(cmd).failed is True:
+        return False
+
+    cmd = "rm -rf /data/web_static/releases/{}/web_static".format(unzipped)
+    if run(cmd).failed is True:
+        return False
+
+    # Relink deployed code
+    if run("rm -rf /data/web_static/current").failed is True:
+        return False
+
+    cmd = "ln -sf /data/web_static/releases/{}/".format(unzipped)
+    cmd = cmd + " /data/web_static/current"
+    if run(cmd).failed is True:
         return False
 
     # Confirmation
-    # print("New version deployed!")
+    print("New version deployed!")
     return True
